+++ app/models/annotation/nml/NmlService.scala
@@ -9,7 +9,8 @@ import com.scalableminds.webknossos.tracingstore.VolumeTracing.VolumeTracing
 import com.typesafe.scalalogging.LazyLogging
 import javax.inject.Inject
 import models.annotation.nml.NmlResults._
-import net.liftweb.common.{Empty, Failure, Full}
+import net.liftweb.common.{Box, Empty, Failure, Full}
+import play.api.i18n.MessagesProvider
 import play.api.libs.Files.{TemporaryFile, TemporaryFileCreator}
 
 import scala.concurrent.ExecutionContext
@@ -17,18 +18,24 @@ import scala.concurrent.ExecutionContext
 class NmlService @Inject()(temporaryFileCreator: TemporaryFileCreator)(implicit ec: ExecutionContext)
     extends LazyLogging {
 
-  def extractFromNml(file: File, name: String): NmlParseResult =
+  def extractFromNml(file: File, name: String)(implicit m: MessagesProvider): NmlParseResult =
     extractFromNml(new FileInputStream(file), name)
 
-  def extractFromNml(inputStream: InputStream, name: String): NmlParseResult =
+  private def formatChain(chain: Box[Failure]): String = chain match {
+    case Full(failure) =>
+      " <~ " + failure.msg + formatChain(failure.chain)
+    case _ => ""
+  }
+
+  def extractFromNml(inputStream: InputStream, name: String)(implicit m: MessagesProvider): NmlParseResult =
     NmlParser.parse(name, inputStream) match {
       case Full((skeletonTracing, volumeTracingWithDataLocation, description, organizationNameOpt)) =>
         NmlParseSuccess(name, skeletonTracing, volumeTracingWithDataLocation, description, organizationNameOpt)
-      case Failure(msg, _, _) => NmlParseFailure(name, msg)
-      case Empty              => NmlParseEmpty(name)
+      case Failure(msg, _, chain) => NmlParseFailure(name, msg + chain.map(_ => formatChain(chain)).getOrElse(""))
+      case Empty                  => NmlParseEmpty(name)
     }
 
-  def extractFromZip(file: File, zipFileName: Option[String] = None): ZipParseResult = {
+  def extractFromZip(file: File, zipFileName: Option[String] = None)(implicit m: MessagesProvider): ZipParseResult = {
     val name = zipFileName getOrElse file.getName
     var otherFiles = Map.empty[String, TemporaryFile]
     var parseResults = List.empty[NmlParseResult]
@@ -91,12 +98,12 @@ class NmlService @Inject()(temporaryFileCreator: TemporaryFileCreator)(implicit 
     }
   }
 
-  def extractFromFiles(files: Seq[(File, String)]): ZipParseResult =
+  def extractFromFiles(files: Seq[(File, String)])(implicit m: MessagesProvider): ZipParseResult =
     files.foldLeft(NmlResults.ZipParseResult()) {
       case (acc, next) => acc.combineWith(extractFromFile(next._1, next._2))
     }
 
-  def extractFromFile(file: File, fileName: String): ZipParseResult =
+  def extractFromFile(file: File, fileName: String)(implicit m: MessagesProvider): ZipParseResult =
     if (fileName.endsWith(".zip")) {
       logger.trace("Extracting from Zip file")
       extractFromZip(file, Some(fileName))