+++ frontend/javascripts/test/shaders/shader_syntax.spec.js
@@ -13,6 +13,7 @@ test("Shader syntax: Ortho Mode", t => {
   const code = getMainFragmentShader({
     colorLayerNames: ["color_layer_1", "color_layer_2"],
     isRgbLayerLookup: { color_layer_1: false, color_layer_2: false },
+    floatLayerLookup: { color_layer_1: false, color_layer_2: false },
     hasSegmentation: false,
     segmentationName: "",
     segmentationPackingDegree: 1,
@@ -34,6 +35,7 @@ test("Shader syntax: Ortho Mode + Segmentation - Mapping", t => {
   const code = getMainFragmentShader({
     colorLayerNames: ["color_layer_1", "color_layer_2"],
     isRgbLayerLookup: { color_layer_1: false, color_layer_2: false },
+    floatLayerLookup: { color_layer_1: false, color_layer_2: false },
     hasSegmentation: true,
     segmentationName: "segmentationLayer",
     segmentationPackingDegree: 1,
@@ -55,6 +57,7 @@ test("Shader syntax: Ortho Mode + Segmentation + Mapping", t => {
   const code = getMainFragmentShader({
     colorLayerNames: ["color_layer_1", "color_layer_2"],
     isRgbLayerLookup: { color_layer_1: false, color_layer_2: false },
+    floatLayerLookup: { color_layer_1: false, color_layer_2: false },
     hasSegmentation: true,
     segmentationName: "segmentationLayer",
     segmentationPackingDegree: 1,
@@ -76,6 +79,7 @@ test("Shader syntax: Arbitrary Mode (no segmentation available)", t => {
   const code = getMainFragmentShader({
     colorLayerNames: ["color_layer_1", "color_layer_2"],
     isRgbLayerLookup: { color_layer_1: false, color_layer_2: false },
+    floatLayerLookup: { color_layer_1: false, color_layer_2: false },
     hasSegmentation: false,
     segmentationName: "",
     segmentationPackingDegree: 1,
@@ -97,6 +101,7 @@ test("Shader syntax: Arbitrary Mode (segmentation available)", t => {
   const code = getMainFragmentShader({
     colorLayerNames: ["color_layer_1", "color_layer_2"],
     isRgbLayerLookup: { color_layer_1: false, color_layer_2: false },
+    floatLayerLookup: { color_layer_1: false, color_layer_2: false },
     hasSegmentation: true,
     segmentationName: "segmentationLayer",
     segmentationPackingDegree: 1,
@@ -105,6 +110,28 @@ test("Shader syntax: Arbitrary Mode (segmentation available)", t => {
     resolutions,
     datasetScale: [1, 1, 1],
     isOrthogonal: false,
+    lookupTextureWidth: DEFAULT_LOOK_UP_TEXTURE_WIDTH,
+  });
+
+  const parseResult = glslParser.check(code);
+
+  t.is(parseResult.log.warningCount, 0);
+  t.is(parseResult.log.errorCount, 0);
+});
+
+test("Shader syntax: Ortho Mode (rgb and float layer)", t => {
+  const code = getMainFragmentShader({
+    colorLayerNames: ["color_layer_1", "color_layer_2"],
+    isRgbLayerLookup: { color_layer_1: true, color_layer_2: false },
+    floatLayerLookup: { color_layer_1: false, color_layer_2: { min: 0, max: 255 } },
+    hasSegmentation: false,
+    segmentationName: "",
+    segmentationPackingDegree: 1,
+    isMappingSupported: true,
+    dataTextureCountPerLayer: 3,
+    resolutions,
+    datasetScale: [1, 1, 1],
+    isOrthogonal: true,
     lookupTextureWidth: DEFAULT_LOOK_UP_TEXTURE_WIDTH,
   });